QR_code_generator/requirements.txt
Pillow==10.3.0
qrcode==7.4.2
streamlit==1.35.0
